[ID: P1-TYPE-ALIAS-SUPPORT-01] C++ 生成コード再生成（py_to inline 化・type alias 反映）

+++ sample/cpp/03_julia_set.cpp
@@ -11,15 +11,15 @@
 
 bytearray render_julia(int64 width, int64 height, int64 max_iter, float64 cx, float64 cy) {
     bytearray pixels = bytearray{};
-    float64 __hoisted_cast_1 = static_cast<float64>(height - 1);
-    float64 __hoisted_cast_2 = static_cast<float64>(width - 1);
-    float64 __hoisted_cast_3 = static_cast<float64>(max_iter);
+    float64 __hoisted_cast_1 = float64(height - 1);
+    float64 __hoisted_cast_2 = float64(width - 1);
+    float64 __hoisted_cast_3 = float64(max_iter);
     
     for (int64 y = 0; y < height; ++y) {
-        float64 zy0 = -(1.2) + 2.4 * (static_cast<float64>(y) / __hoisted_cast_1);
+        float64 zy0 = -(1.2) + 2.4 * (float64(y) / __hoisted_cast_1);
         
         for (int64 x = 0; x < width; ++x) {
-            float64 zx = -(1.8) + 3.6 * (static_cast<float64>(x) / __hoisted_cast_2);
+            float64 zx = -(1.8) + 3.6 * (float64(x) / __hoisted_cast_2);
             float64 zy = zy0;
             
             int64 i = 0;
@@ -40,14 +40,14 @@ bytearray render_julia(int64 width, int64 height, int64 max_iter, float64 cx, fl
                 g = 0;
                 b = 0;
             } else {
-                float64 t = static_cast<float64>(i) / __hoisted_cast_3;
-                r = static_cast<int64>(255.0 * (0.2 + 0.8 * t));
-                g = static_cast<int64>(255.0 * (0.1 + 0.9 * t * t));
-                b = static_cast<int64>(255.0 * (1.0 - t));
+                float64 t = float64(i) / __hoisted_cast_3;
+                r = int64(255.0 * (0.2 + 0.8 * t));
+                g = int64(255.0 * (0.1 + 0.9 * t * t));
+                b = int64(255.0 * (1.0 - t));
             }
-            pixels.append(static_cast<uint8>(static_cast<int64>(r)));
-            pixels.append(static_cast<uint8>(static_cast<int64>(g)));
-            pixels.append(static_cast<uint8>(static_cast<int64>(b)));
+            pixels.append(uint8(r));
+            pixels.append(uint8(g));
+            pixels.append(uint8(b));
         }
     }
     return pixels;
